+++ spec/dummy/app/models/application_record.rb
@@ -9,5 +9,6 @@ class ApplicationRecord < ActiveRecord::Base
   include Agilibox::PluckToHash
   include Agilibox::PolymorphicId
   include Agilibox::Search
+  include Agilibox::TapMethods
   include Agilibox::TimestampHelpers
 end