+++ CHANGELOG.md
@@ -3,6 +3,13 @@
 ## Next version
 - Remove `GetHTTP`
 - I18n fixes
+- Add `Agilibox::TapMethods`
+- Filters CSS : add flex-wrap
+- Add `bs_card` helper
+- Add `:br` separator to `info` helper
+- TextHelper refactors
+- Update dummy app to Rails 5.2
+- Replace AXLSX by SpreadsheetArchitect
 
 ## 1.5.13
 - Fix `info` helper with nested blank value